Move middleware and auth under ewb
