+++ src/web/articles/favorite.rs
@@ -1,6 +1,6 @@
 use crate::domain::repositories::Repository;
-use crate::middleware::ContextExt;
 use crate::web::articles::responses::ArticleResponse;
+use crate::web::middleware::ContextExt;
 use crate::Context;
 use tide::{Request, Response};
