+++ src/web/articles/insert.rs
@@ -1,6 +1,6 @@
 use crate::domain::repositories::Repository;
-use crate::middleware::ContextExt;
 use crate::web::articles::responses::ArticleResponse;
+use crate::web::middleware::ContextExt;
 use crate::{domain, Context};
 use serde::{Deserialize, Serialize};
 use tide::Response;
