+++ src/web/articles/list.rs
@@ -1,7 +1,7 @@
 use crate::domain::articles::ArticleQuery;
 use crate::domain::repositories::Repository;
-use crate::middleware::ContextExt;
 use crate::web::articles::responses::ArticlesResponse;
+use crate::web::middleware::ContextExt;
 use crate::Context;
 use tide::{Request, Response};
 use uuid::Uuid;
