+++ src/web/auth.rs
