+++ src/web/comments/create.rs
@@ -1,7 +1,7 @@
 use crate::domain::repositories::Repository;
 use crate::domain::CommentContent;
-use crate::middleware::ContextExt;
 use crate::web::comments::responses::CommentResponse;
+use crate::web::middleware::ContextExt;
 use crate::Context;
 use serde::{Deserialize, Serialize};
 use tide::Response;
