+++ src/web/comments/get.rs
@@ -1,6 +1,6 @@
 use crate::domain::repositories::Repository;
-use crate::middleware::ContextExt;
 use crate::web::comments::responses::CommentsResponse;
+use crate::web::middleware::ContextExt;
 use crate::Context;
 use tide::Response;
 use uuid::Uuid;
