+++ tests/web_comments.rs
@@ -7,7 +7,7 @@ use helpers::{create_article, create_user, create_users};
 
 use async_std::task;
 use fake::fake;
-use realworld_tide::auth::encode_token;
+use realworld_tide::web::auth::encode_token;
 use realworld_tide::web::comments::create::NewCommentRequest;
 
 #[test]
